ActiveRecordCoder: encode wether the record was persisted or not
Fix: https://github.com/Shopify/paquito/issues/10
Closes: https://github.com/Shopify/paquito/pull/11

Unfortunately `AR::Base.instantiate` doesn't provide a way to pass
wether the record was persisted or not, so we have to bypass it.

By doing so we skip `discriminate_class_for_record` which decide which
subclass to use for STI, but I don't think we need it.

Important Note: this change is able to read payload generated by older
versions of Paquito, however *older versions of Paquito will fail to
read these new payloads*