Introduce new packers for Time and DateTime
Fix: https://github.com/Shopify/paquito/issues/26

We use an uint32 to store `time.to_r.denominator`, but this is
incorrect as the denominator can be of any size.

Additionally we serialize it with `Array#pack` which silently ignore
bits out of the range: https://bugs.ruby-lang.org/issues/19245

```ruby
>> [2**32].pack("L>").unpack("L>")
=> [0]
>> [2**32 + 12].pack("L>").unpack("L>")
=> [12]
```

Another issue I noticed is that we always store Time instance in
UTC. This is generally what you want, but I think is a dangerious
behavior and we'd be better to be as precise as possible.

We could persist the `utc_offset` in msgpack, however after investigating
I couldn't find a way to restore the `#zone` attribute. None of the
`Time` constructors allow this.

So I decided to use `Marshal` for `Time` and `DateTime`, as it's the
fastest precise encoding method I could find.

To allow for a safe rollout however, we need to allow to first deploy
with the capability to read the older formats but without writing them.
For that purpose we introduce `Paquito.format_version` which control which
version of the serializer is used for writing.