Introduce new packers for Time and DateTime
Fix: https://github.com/Shopify/paquito/issues/26

We use an uint32 to store `time.to_r.denominator`, but this is
incorrect as the denominator can be of any size.

Additionally we serialize it with `Array#pack` which silently ignore
bits out of the range: https://bugs.ruby-lang.org/issues/19245

```ruby
>> [2**32].pack("L>").unpack("L>")
=> [0]
>> [2**32 + 12].pack("L>").unpack("L>")
=> [12]
```

We can solve this by using a recursive serializer, the multiple
integers composing a `Time` instance are now serialized by msgpack
which handle 64bit integers natively, and also optionally handle
arbitrary size big ints.

Another issue I noticed is that we always store Time instance in
UTC. This is generally what you want, but I think is a dangerous
behavior and we'd be better to be as precise as possible.

So we now persist and restore the `utc_offset` in msgpack.
However after investigating I couldn't find a way to restore the `#zone`
attribute. None of the `Time` constructors allow this.
But I think it's OK for now as this attribute is never really used
and not really usable either.

To allow for a safe rollout however, we need to allow to first deploy
with the capability to read the older formats but without writing them.
For that purpose we introduce `Paquito.format_version` which control which
version of the serializer is used for writing.